My solo updates on 10/20
I finished check_valid_move() by fixing the logic of the first if statement as well as typos in the declaration of a for loop in board_has_no_duplicates(). I then re-started solve_three_in_a_row() and solve_three_in_a_column(), as I realized that they're supposed to fill in a square next to two squares of the same color even if there isn't another unknown on the other side.